# src/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Vlad Romaniuk",
    "version": "1.0.0",
    "title": "Flask Swagger UI Tutorial",
    "contact": {
      "email": "romanuk_vlad@ukr.net"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "Film Request",
      "description": "Example API for requesting and return films requests"
    }
  ],
  "paths": {
    "/films": {
      "get": {
        "tags": [
          "Film Request"
        ],
        "summary": "Returns all available films in the system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/filmsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Film Request"
        ],
        "summary": "Create a new film",
        "requestBody": {
          "description": "Film Request Post Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/films"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          }
        }
      }
    },
    "/films/{uuid}": {
      "get": {
        "tags": [
          "Film Request"
        ],
        "summary": "Get film with given UUID",
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Film Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/film"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          },
          "404": {
            "description": "Failed. Film was not found."
          }
        }
      },
      "put": {
        "summary": "edit a film request by uuid",
        "tags": [
          "Film Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Film Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Film Request Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/film"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/film"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          }
        }
      },
      "patch": {
        "summary": "edit a film parameter request by uuid",
        "tags": ["Film Request"],
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Film Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Film Request Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/film"
              }
            }
          }
        },
      "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/film"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          }
        }
      },
      "delete": {
        "summary": "Delete Film Request by UUID",
        "tags": [
          "Film Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Film Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          },
          "400": {
            "description": "Failed. Misunderstood Request."
          },
          "404": {
            "description": "Failed. Film was not found."
          }
        }
      }
    },
    "/actors": {
      "get": {
        "tags": [
          "Actor Request"
        ],
        "summary": "Returns all available actors in the system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/actorsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Actor Request"
        ],
        "summary": "Create a new actor",
        "requestBody": {
          "description": "Actor Request Post Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/actors"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          }
        }
      }
    },
    "/actors/{uuid}":{
      "get": {
        "tags": [
          "Actor Request"
        ],
        "summary": "Get actor with given UUID",
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Actor Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/actors"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          },
          "404": {
            "description": "Failed. Actor was not found."
          }
        }
      },
      "put": {
        "summary": "edit an actor request by uuid",
        "tags": [
          "Actor Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Actor Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Actor Request Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/actors"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/actors"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          }
        }
      },
      "patch": {
        "summary": "edit an actor parameter request by uuid",
        "tags": ["Actor Request"],
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Actor Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Actor Request Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/actors"
              }
            }
          }
        },
      "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/actors"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          }
        }
      },
      "delete": {
        "summary": "Delete Actor Request by UUID",
        "tags": [
          "Actor Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Actor Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          },
          "400": {
            "description": "Failed. Misunderstood Request."
          },
          "404": {
            "description": "Failed. Actor was not found."
          }
        }
      }
    },
    "/register":{
      "post": {
        "tags": [
          "Register"
        ],
        "summary": "Create a new user",
        "requestBody": {
          "description": "User Request Post Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          }
        }
      }
    }
  }
}